fixed some tests, starting to make docs
